Eagerly read all i18n YAML files, add more tests
This makes `New()` read all the files it found eagerly on instantiation.
The benefit of doing this lazily is negligible – especially when I look
at the benchmark and the tests which immediately load every single file
after instantiation anyway. Changing the backend’s data structure to a
slice of byte slices will become clear in the next commmit.